requirements.txt
Django~=4.2.2
djangorestframework~=3.14.0
setuptools~=65.5.1